# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for NOVA Skin Med Spa public site and administration.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: number;
  email: string;
  name: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface SiteSettings {
  id: number;
  clinicName: string;
  tagline: string;
  phone: string;
  whatsapp: string;
  email: string;
  address: string;
  hours: string;
  instagram: string;
  facebook: string;
  tiktok: string;
  heroImage: string;
  heroEyebrow: string;
  heroTitle: string;
  heroDescription: string;
  aboutText: string;
  updatedAt: string;
}

export interface SiteSettingsInput {
  /** @minLength 1 */
  clinicName: string;
  tagline: string;
  phone: string;
  whatsapp: string;
  email: string;
  address: string;
  hours: string;
  instagram: string;
  facebook: string;
  tiktok: string;
  heroImage: string;
  heroEyebrow: string;
  heroTitle: string;
  heroDescription: string;
  aboutText: string;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  /** @nullable */
  price: number | null;
  /** @nullable */
  duration: string | null;
  imageUrl: string;
  active: boolean;
  sortOrder: number;
  createdAt: string;
}

export interface ServiceInput {
  /** @minLength 1 */
  name: string;
  description: string;
  /** @nullable */
  price: number | null;
  /** @nullable */
  duration: string | null;
  imageUrl: string;
  active: boolean;
  sortOrder: number;
}

export interface GalleryImage {
  id: number;
  title: string;
  description: string;
  imageUrl: string;
  active: boolean;
  sortOrder: number;
  createdAt: string;
}

export interface GalleryImageInput {
  /** @minLength 1 */
  title: string;
  description: string;
  imageUrl: string;
  active: boolean;
  sortOrder: number;
}

export interface Video {
  id: number;
  title: string;
  description: string;
  videoUrl: string;
  /** @nullable */
  posterUrl: string | null;
  active: boolean;
  sortOrder: number;
  createdAt: string;
}

export interface VideoInput {
  /** @minLength 1 */
  title: string;
  description: string;
  videoUrl: string;
  /** @nullable */
  posterUrl: string | null;
  active: boolean;
  sortOrder: number;
}

export interface Promotion {
  id: number;
  title: string;
  description: string;
  imageUrl: string;
  startDate: string;
  endDate: string;
  active: boolean;
  createdAt: string;
}

export interface PromotionInput {
  /** @minLength 1 */
  title: string;
  description: string;
  imageUrl: string;
  startDate: string;
  endDate: string;
  active: boolean;
}

export interface Specialist {
  id: number;
  name: string;
  specialty: string;
  bio: string;
  photoUrl: string;
  /** @nullable */
  instagram: string | null;
  active: boolean;
  createdAt: string;
}

export interface SpecialistInput {
  /** @minLength 1 */
  name: string;
  specialty: string;
  bio: string;
  photoUrl: string;
  /** @nullable */
  instagram: string | null;
  active: boolean;
}

export interface Testimonial {
  id: number;
  name: string;
  comment: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  photoUrl: string | null;
  active: boolean;
  createdAt: string;
}

export interface TestimonialInput {
  /** @minLength 1 */
  name: string;
  comment: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  photoUrl: string | null;
  active: boolean;
}

export interface ContactMessage {
  id: number;
  name: string;
  phone: string;
  email: string;
  message: string;
  read: boolean;
  createdAt: string;
}

export interface ContactMessageInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  email: string;
  /** @minLength 1 */
  message: string;
}

export interface MessageUpdate {
  read: boolean;
}

export interface AdminSummary {
  images: number;
  videos: number;
  services: number;
  promotions: number;
  messages: number;
  unreadMessages: number;
}

export interface SiteContent {
  settings: SiteSettings;
  services: Service[];
  gallery: GalleryImage[];
  videos: Video[];
  promotions: Promotion[];
  specialists: Specialist[];
  testimonials: Testimonial[];
}