# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API for NOVA Skin Med Spa public site and administration.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contact:
    post:
      operationId: createContactMessage
      tags: [public]
      summary: Submit a contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactMessageInput"
      responses:
        "201":
          description: Message received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Start an administrator session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the current administrator
      responses:
        "200":
          description: Current administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
...
  /admin/settings:
    get:
      operationId: getSiteSettings
      tags: [admin]
      summary: Get editable site settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
    put:
      operationId: updateSiteSettings
      tags: [admin]
      summary: Update editable site settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
  /admin/services:
    get:
      operationId: listServices
      tags: [admin]
      summary: List services
      responses:
        "200":
          description: Services
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Service" }
    post:
      operationId: createService
      tags: [admin]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ServiceInput" }
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Service" }
...
  /admin/gallery:
    get:
      operationId: listGalleryImages
      tags: [admin]
      summary: List gallery images
      responses:
        "200":
          description: Gallery images
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GalleryImage" }
    post:
      operationId: createGalleryImage
      tags: [admin]
      summary: Create gallery image
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GalleryImageInput" }
      responses:
        "201":
          description: Created gallery image
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GalleryImage" }
...
  /admin/videos:
    get:
      operationId: listVideos
      tags: [admin]
      summary: List videos
      responses:
        "200":
          description: Videos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Video" }
    post:
      operationId: createVideo
      tags: [admin]
      summary: Create video
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoInput" }
      responses:
        "201":
          description: Created video
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Video" }
...
  /admin/promotions:
    get:
      operationId: listPromotions
      tags: [admin]
      summary: List promotions
      responses:
        "200":
          description: Promotions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Promotion" }
    post:
      operationId: createPromotion
      tags: [admin]
      summary: Create promotion
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PromotionInput" }
      responses:
        "201":
          description: Created promotion
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Promotion" }
...
  /admin/specialists:
    get:
      operationId: listSpecialists
      tags: [admin]
      summary: List specialists
      responses:
        "200":
          description: Specialists
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Specialist" }
    post:
      operationId: createSpecialist
      tags: [admin]
      summary: Create specialist
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SpecialistInput" }
      responses:
        "201":
          description: Created specialist
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Specialist" }
...
  /admin/testimonials:
    get:
      operationId: listTestimonials
      tags: [admin]
      summary: List testimonials
      responses:
        "200":
          description: Testimonials
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Testimonial" }
    post:
      operationId: createTestimonial
      tags: [admin]
      summary: Create testimonial
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TestimonialInput" }
      responses:
        "201":
          description: Created testimonial
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Testimonial" }
...
  /admin/messages/{id}:
    parameters:
      - { $ref: "#/components/parameters/Id" }
    patch:
      operationId: updateContactMessage
      tags: [admin]
      summary: Mark a message as read or unread
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MessageUpdate" }
      responses:
        "200":
          description: Updated message
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContactMessage" }
    delete:
      operationId: deleteContactMessage
      tags: [admin]
      summary: Delete contact message
      responses:
        "204": { description: Deleted }
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer }
  schemas:
...
    AuthUser:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
      required: [id, email, name]
    LoginInput:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
      required: [email, password]
    SiteSettings:
      type: object
      properties:
        id: { type: integer }
        clinicName: { type: string }
        tagline: { type: string }
        phone: { type: string }
        whatsapp: { type: string }
        email: { type: string }
        address: { type: string }
        hours: { type: string }
        instagram: { type: string }
        facebook: { type: string }
        tiktok: { type: string }
        heroImage: { type: string }
        heroEyebrow: { type: string }
        heroTitle: { type: string }
        heroDescription: { type: string }
        aboutText: { type: string }
        updatedAt: { type: string, format: date-time }
      required: [id, clinicName, tagline, phone, whatsapp, email, address, hours, instagram, facebook, tiktok, heroImage, heroEyebrow, heroTitle, heroDescription, aboutText, updatedAt]
    SiteSettingsInput:
      type: object
      properties:
        clinicName: { type: string, minLength: 1 }
        tagline: { type: string }
        phone: { type: string }
        whatsapp: { type: string }
        email: { type: string, format: email }
        address: { type: string }
        hours: { type: string }
        instagram: { type: string }
        facebook: { type: string }
        tiktok: { type: string }
        heroImage: { type: string }
        heroEyebrow: { type: string }
        heroTitle: { type: string }
        heroDescription: { type: string }
        aboutText: { type: string }
      required: [clinicName, tagline, phone, whatsapp, email, address, hours, instagram, facebook, tiktok, heroImage, heroEyebrow, heroTitle, heroDescription, aboutText]
    Service:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        price: { type: ["number", "null"] }
        duration: { type: ["string", "null"] }
        imageUrl: { type: string }
        active: { type: boolean }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, name, description, price, duration, imageUrl, active, sortOrder, createdAt]
    ServiceInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        price: { type: ["number", "null"] }
        duration: { type: ["string", "null"] }
        imageUrl: { type: string }
        active: { type: boolean }
        sortOrder: { type: integer }
      required: [name, description, price, duration, imageUrl, active, sortOrder]
    GalleryImage:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        imageUrl: { type: string }
        active: { type: boolean }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, title, description, imageUrl, active, sortOrder, createdAt]
    GalleryImageInput:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        imageUrl: { type: string }
        active: { type: boolean }
        sortOrder: { type: integer }
      required: [title, description, imageUrl, active, sortOrder]
    Video:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        videoUrl: { type: string }
        posterUrl: { type: ["string", "null"] }
        active: { type: boolean }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, title, description, videoUrl, posterUrl, active, sortOrder, createdAt]
    VideoInput:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        videoUrl: { type: string }
        posterUrl: { type: ["string", "null"] }
        active: { type: boolean }
        sortOrder: { type: integer }
      required: [title, description, videoUrl, posterUrl, active, sortOrder]
    Promotion:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        imageUrl: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, title, description, imageUrl, startDate, endDate, active, createdAt]
    PromotionInput:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        imageUrl: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        active: { type: boolean }
      required: [title, description, imageUrl, startDate, endDate, active]
    Specialist:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        specialty: { type: string }
        bio: { type: string }
        photoUrl: { type: string }
        instagram: { type: ["string", "null"] }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, name, specialty, bio, photoUrl, instagram, active, createdAt]
    SpecialistInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        specialty: { type: string }
        bio: { type: string }
        photoUrl: { type: string }
        instagram: { type: ["string", "null"] }
        active: { type: boolean }
      required: [name, specialty, bio, photoUrl, instagram, active]
    Testimonial:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        comment: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
        photoUrl: { type: ["string", "null"] }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, name, comment, rating, photoUrl, active, createdAt]
    TestimonialInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        comment: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
        photoUrl: { type: ["string", "null"] }
        active: { type: boolean }
      required: [name, comment, rating, photoUrl, active]
    ContactMessage:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        phone: { type: string }
        email: { type: string }
        message: { type: string }
        read: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, name, phone, email, message, read, createdAt]
    ContactMessageInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string, minLength: 1 }
        email: { type: string, format: email }
        message: { type: string, minLength: 1 }
      required: [name, phone, email, message]
    MessageUpdate:
      type: object
      properties:
        read: { type: boolean }
      required: [read]
...